Add files via upload
An app is also developed which is using MITAppInventor, which is an online software. The rover movement and all is fully controlled using this app. One servo motor is used, which is attached with the moisture sensor because while the servo rotates the moisture sensor will dip in the soil in order to get the moisture content in the soil.